fix: add rainbowkit mock to ViewApplicationPage.test.tsx and move into __tests__
